requirements.txt
pyzmq==25.1.1
aiohttp==3.8.5
numpy==1.24.3
websocket-client==1.6.3
